修复部分bug（指更新ui写重复了的方法

+++ build.bat
@@ -1,29 +1,28 @@
 @echo off
 echo [1/3] Cleaning old build files...
-rd /s /q build dist
-del /q *.spec
+if exist build rd /s /q build
+if exist dist rd /s /q dist
+if exist *.spec del /q *.spec
 
-echo [2/3] Starting PyInstaller (Bundling internal tools)...
+echo [2/3] Starting PyInstaller...
+:: 使用 --collect-all 代替手动寻找路径，这更安全且不容易出错
 pyinstaller --noconfirm --onefile --windowed ^
 --add-data "wowsunpack.exe;." ^
 --add-data "pfsunpack.exe;." ^
 --add-data "GameParams.py;." ^
 --hidden-import="GameParams" ^
+--collect-all "customtkinter" ^
 --name "WowsAnalyzer" ^
 "MainUI.py"
 
-:: Check if the previous command failed
 if %ERRORLEVEL% NEQ 0 (
     echo.
-    echo [! ERROR] Build failed. Please check the logs above.
+    echo [! ERROR] Build failed.
     pause
     exit /b %ERRORLEVEL%
 )
 
 echo [3/3] Organizing distribution folder...
-cd dist
-
-echo.
-echo Build Successful! Closing in 3 seconds...
+echo Build Successful!
 timeout /t 3
 exit
